pyproject.toml
[project]
name = "polymorph"
version = "1.0.0"
description = "Federated Learning Backdoor Framework with PolyMorph Attack"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "PyYAML>=6.0",
    "Pillow>=9.0.0",
    "scikit-learn>=1.0.0",
    "hdbscan>=0.8.0",
    "tqdm>=4.60.0",
    "wandb>=0.15.0",
    "psutil>=5.8.0",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.0.0",
    "ipykernel>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]
